removes redundant inherits structure
