+++ DACMASDK-Sample/src/main/res/layout/fragment_content_ad.xml
@@ -1,0 +1,56 @@
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent" >
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <!-- コンテンツ動画View -->
+        <jp.co.dac.ma.sdk.widget.DACVideoPlayerView
+            android:id="@+id/content_video_player"
+            android:layout_width="match_parent"
+            android:layout_height="240dp"
+            android:layout_gravity="center"
+            android:visibility="gone" />
+
+        <!-- 広告レイアウト -->
+        <include
+            android:id="@+id/dac_ad_layout"
+            layout="@layout/dac_ad_layout" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="150dp"
+            android:background="@android:color/black"
+            android:textColor="@android:color/white"
+            android:text="test1" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="100dp"
+            android:background="@android:color/holo_blue_bright"
+            android:text="test2" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="300dp"
+            android:background="@android:color/holo_green_dark"
+            android:text="test3" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="500dp"
+            android:background="@android:color/holo_orange_dark"
+            android:text="test4" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="500dp"
+            android:background="@android:color/holo_green_dark"
+            android:text="test5" />
+    </LinearLayout>
+
+</ScrollView>
