--- DACMASDK-Sample/src/main/res/layout/fragment_content_video_player.xml
@@ -1,124 +1,0 @@
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent" >
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <!-- for content -->
-        <jp.co.dac.ma.sdk.widget.DACVideoPlayerView
-            android:id="@+id/content_video_player"
-            android:layout_width="match_parent"
-            android:layout_height="240dp"
-            android:layout_gravity="center"
-            android:visibility="gone" />
-
-        <!-- ここから広告 -->
-        <jp.co.dac.sdk.ma.sample.VideoPlayerContentWithAdPlayback
-            android:id="@+id/video_player_with_ad_playback"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <FrameLayout
-                android:id="@+id/video_player_image"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_centerHorizontal="true"
-                android:visibility="gone" />
-
-            <jp.co.dac.ma.sdk.widget.DACVideoPlayerView
-                android:id="@+id/ad_video_player"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_centerHorizontal="true" />
-
-            <RelativeLayout
-                android:id="@+id/video_controller"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:padding="6dp" >
-
-                <jp.co.dac.ma.sdk.widget.MuteButton
-                    android:id="@+id/mute_button"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:scaleType="fitCenter"
-                    android:background="@android:color/transparent" />
-
-                <jp.co.dac.ma.sdk.widget.PRButton
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:scaleType="fitCenter"
-                    android:layout_alignParentRight="true"
-                    android:background="@android:color/transparent" />
-
-                <jp.co.dac.ma.sdk.widget.FullscreenButton
-                    android:id="@+id/fullscreen_button"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_alignParentRight="true"
-                    android:layout_alignParentBottom="true"
-                    android:scaleType="fitCenter"
-                    android:background="@android:color/transparent" />
-
-                <jp.co.dac.ma.sdk.widget.ReplayButton
-                    android:id="@+id/replay_button"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_alignParentBottom="true"
-                    android:scaleType="fitCenter"
-                    android:visibility="gone"
-                    android:background="@android:color/transparent" />
-
-                <jp.co.dac.ma.sdk.widget.SkipButton
-                    android:id="@+id/skip_button"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_alignParentRight="true"
-                    android:layout_centerVertical="true"
-                    android:scaleType="fitCenter"
-                    android:visibility="gone"
-                    android:src="@android:drawable/ic_media_next"
-                    android:background="@android:color/transparent" />
-            </RelativeLayout>
-
-        </jp.co.dac.sdk.ma.sample.VideoPlayerContentWithAdPlayback>
-        <!-- 広告終了 -->
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="150dp"
-            android:background="@android:color/black"
-            android:textColor="@android:color/white"
-            android:text="test1" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="100dp"
-            android:background="@android:color/holo_blue_bright"
-            android:text="test2" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="300dp"
-            android:background="@android:color/holo_green_dark"
-            android:text="test3" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="500dp"
-            android:background="@android:color/holo_orange_dark"
-            android:text="test4" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="500dp"
-            android:background="@android:color/holo_green_dark"
-            android:text="test5" />
-    </LinearLayout>
-
-</ScrollView>
