+++ DACMASDK-Sample/src/noContent/java/jp/co/dac/sdk/ma/sample/AdFragment.java
@@ -30,7 +30,7 @@ public class AdFragment extends Fragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_scrollable_video_player, container, false);
+        return inflater.inflate(R.layout.fragment_ad, container, false);
     }
 
     @Override
