+++ Makefile
@@ -1,7 +1,7 @@
 all_test: dac_ma_test dac_brightcove_test
 
 dac_ma_test: # run DACMASDK test
-	gradle :DACMASDK-Sample:spoon && open DACMASDK-Sample/build/spoon/debug/index.html
+	gradle :DACMASDK-Sample:spoon
 
 dac_brightcove_test: # run Brightcove test(now, experimental)
-	gradle :Brightcove-Sample:spoon && open Brightcove-Sample/build/spoon/debug/index.html
+	gradle :Brightcove-Sample:spoon
